pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.38",
    "werkzeug>=3.1.3",
    "trafilatura>=2.0.0",
    "flask-migrate>=4.1.0",
    "python-dotenv>=1.0.1",
    "mysql-connector-python>=9.2.0",
    "pymysql>=1.1.1",
]
